Feat admin activity

+++ app/src/main/java/com/byoon/lastminuteflix/db/UserDao.java
@@ -27,7 +27,7 @@ public interface UserDao {
   void delete(User... users);
 
   @Query("DELETE FROM " + AppDatabase.USER_TABLE + " WHERE mUserId = :userId")
-  void deleteUserById(int userId);
+  void deleteUserById(long userId);
 
   @Query("SELECT * FROM " + AppDatabase.USER_TABLE)
   List<User> getAllUsers();
